Move target resolving from cli to api
